// typings/contracts/voting/testing/ExposedBallot.ts
export declare namespace Ballot {
  export type VoteStruct = { choice: BytesLike; hasVoted: boolean };

  export type VoteStructOutput = [string, boolean] & {
    choice: string;
    hasVoted: boolean;
  };
}

export declare namespace ExposedBallot {
  export type ResultSampleStruct = { voter: string; vote: Ballot.VoteStruct };

  export type ResultSampleStructOutput = [string, Ballot.VoteStructOutput] & {
    voter: string;
    vote: Ballot.VoteStructOutput;
  };
}